lc0120: Triangle
