+++ src/main/java/com/zea7ot/lc/logs.txt
@@ -1,3 +1,8 @@
+20200714:
+3: 
+4: 0120
+5: 
+
 20200713:
 3: 0200, 0763, 0994
 4: 0018, 0146, 1192
