+++ src/main/java/com/zea7ot/lc/lvl4/lc0120/SolutionApproach0DP2.java
@@ -5,19 +5,9 @@ package com.zea7ot.lc.lvl4.lc0120;
 
 import java.util.List;
 
-public class SolutionApproachDPInPlace {
-    public int minimumTotal(List<List<Integer>> triangle) {
-        for(int i = triangle.size() - 2; i >= 0; i--){
-            for(int j = 0; j <= i; j++){
-                triangle.get(i).set(j, triangle.get(i).get(j) + Math.min(triangle.get(i + 1).get(j), triangle.get(i + 1).get(j + 1)));
-            }
-        }
-        
-        return triangle.get(0).get(0);
-    }
-
+public class SolutionApproach0DP2 {
     // to elaborate the above method
-    public int minimumTotal2(List<List<Integer>> triangle) {
+    public int minimumTotal(List<List<Integer>> triangle) {
         for(int i = triangle.size() - 2; i >= 0; i--){
             for(int j = 0; j <= i; j++){
                 int self = triangle.get(i).get(j);  // to get the [row](i + 1) & [col](j + 1)
