--- src/main/java/com/zea7ot/lc/lvl4/lc0120/SolutionApproachDP.java
@@ -1,23 +1,0 @@
-/**
- * https://leetcode.com/problems/triangle/
- */
-package com.zea7ot.lc.lvl4.lc0120;
-
-import java.util.List;
-
-public class SolutionApproachDP {
-    public int minimumTotal(List<List<Integer>> triangle) {
-        final int SIZE = triangle.size();
-        int[] results = new int[SIZE + 1];
-        
-        for(int i = SIZE - 1; i >= 0; i--){
-            List<Integer> temp = triangle.get(i);
-            
-            for(int j = 0; j < temp.size(); j++){
-                results[j] = Math.min(results[j], results[j + 1]) + temp.get(j);
-            }
-        }
-        
-        return results[0];
-    }
-}
